requirements.txt
flask==3.0.3
flask-cors==4.0.1
openai==1.30.1
anthropic==0.28.0
google-generativeai==0.7.2
requests==2.32.3
beautifulsoup4==4.12.3
pypdf2==3.0.1
pdfplumber==0.11.0
reportlab==4.2.2
pandas==2.2.2
python-dotenv==1.0.1
Werkzeug==3.0.3
httpx>=0.27
